+++ src/main/java/blackjack/view/OutputView.java
@@ -1,8 +1,8 @@
 package blackjack.view;
 
-import blackjack.domain.Card;
-import blackjack.domain.Rank;
-import blackjack.domain.Suit;
+import blackjack.domain.card.Card;
+import blackjack.domain.card.Rank;
+import blackjack.domain.card.Suit;
 import blackjack.domain.Victory;
 import blackjack.domain.WinningResult;
 import blackjack.domain.participants.Dealer;
