+++ src/test/java/blackjack/domain/DeckTest.java
@@ -1,5 +1,9 @@
 package blackjack.domain;
 
+import blackjack.domain.card.Card;
+import blackjack.domain.card.Deck;
+import blackjack.domain.card.Rank;
+import blackjack.domain.card.Suit;
 import java.util.List;
 import java.util.Stack;
 import org.assertj.core.api.Assertions;
