+++ src/test/java/blackjack/domain/FixCardsShuffler.java
@@ -1,5 +1,7 @@
 package blackjack.domain;
 
+import blackjack.domain.card.Card;
+import blackjack.domain.card.CardsShuffler;
 import java.util.List;
 
 public class FixCardsShuffler implements CardsShuffler {
