+++ src/test/java/blackjack/domain/VictoryTest.java
@@ -1,5 +1,11 @@
 package blackjack.domain;
 
+import blackjack.domain.card.Card;
+import blackjack.domain.card.Cards;
+import blackjack.domain.card.Deck;
+import blackjack.domain.card.Rank;
+import blackjack.domain.card.ScoreCalculator;
+import blackjack.domain.card.Suit;
 import blackjack.domain.participants.Dealer;
 import blackjack.domain.participants.Player;
 import blackjack.domain.participants.Players;
